Refactor NetRemapper and update tests/resources
Refactored the `NetRemapper` class to include constructors for both file paths and streams, and moved property initializations out of the constructor. Added a `MappingsLoaded` property to check if mappings are loaded. Updated `MappingsReaderTests.cs` to check for the existence of `./Resources/Main.exe`. Modified `NetRemapper-Tests.csproj` to copy resource files to the output directory. Added `Main.exe` and an XML manifest file for testing and application settings.